Move collect_associated_files() to bids.utils.
